+++ packages/contracts/test/core/permission/permission-manager.ts
@@ -9,7 +9,6 @@ import {
   PermissionConditionMock__factory,
   TestPlugin__factory,
 } from '../../../typechain';
-import {DeployTestPermissionCondition} from '../../test-utils/conditions';
 import {OZ_ERRORS} from '../../test-utils/error';
 import {Operation} from '../../../utils/types';
 
@@ -728,7 +727,7 @@ describe('Core: PermissionManager', function () {
   });
 
   describe('isGranted', () => {
-    it('should return true if the permission is granted to the user', async () => {
+    it('returns `true` if the permission is granted to the user', async () => {
       await pm.grant(pm.address, otherSigner.address, ADMIN_PERMISSION_ID);
       const isGranted = await pm.callStatic.isGranted(
         pm.address,
@@ -739,7 +738,7 @@ describe('Core: PermissionManager', function () {
       expect(isGranted).to.be.equal(true);
     });
 
-    it('should return false if the permissions is not granted to the user', async () => {
+    it('returns `false` if the permission is not granted to the user', async () => {
       const isGranted = await pm.callStatic.isGranted(
         pm.address,
         otherSigner.address,
@@ -749,35 +748,112 @@ describe('Core: PermissionManager', function () {
       expect(isGranted).to.be.equal(false);
     });
 
-    it('should return true for permissions granted to any address on a specific target contract using the `ANY_ADDR` flag', async () => {
-      const anyAddr = await pm.getAnyAddr();
-      const condition = await DeployTestPermissionCondition();
+    it('returns `true` if a condition is set for a specific caller and target answering `true`', async () => {
+      const condition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
       await pm.grantWithCondition(
         pm.address,
-        anyAddr,
+        ownerSigner.address,
         ADMIN_PERMISSION_ID,
         condition.address
       );
+      await condition.setAnswer(true);
+
+      expect(
+        await pm.isGranted(
+          pm.address,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          condition.address
+        )
+      ).to.be.true;
+    });
+
+    it('returns `true` if a condition is set for a generic caller answering `true`', async () => {
+      const condition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
       await pm.grantWithCondition(
-        anyAddr,
         pm.address,
+        ANY_ADDR,
         ADMIN_PERMISSION_ID,
         condition.address
       );
-      const isGranted_1 = await pm.callStatic.isGranted(
-        pm.address,
-        anyAddr,
+      await condition.setAnswer(true);
+
+      // Check `ownerSigner.address` as a caller `_who`
+      expect(
+        await pm.isGranted(
+          pm.address,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          condition.address
+        )
+      ).to.be.true;
+
+      // Check `otherSigner.address` as a caller `_who`
+      expect(
+        await pm.isGranted(
+          pm.address,
+          otherSigner.address,
+          ADMIN_PERMISSION_ID,
+          condition.address
+        )
+      ).to.be.true;
+
+      // Check that `false` is returned if `address(0)` is the target `_where`.
+      expect(
+        await pm.isGranted(
+          ethers.constants.AddressZero,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          condition.address
+        )
+      ).to.be.false;
+    });
+
+    it('returns `true` if a condition is set for a generic target answering `true`', async () => {
+      const condition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+      await pm.grantWithCondition(
+        ANY_ADDR,
+        ownerSigner.address,
         ADMIN_PERMISSION_ID,
         condition.address
       );
-      const isGranted_2 = await pm.callStatic.isGranted(
-        pm.address,
-        anyAddr,
-        ADMIN_PERMISSION_ID,
-        condition.address
-      );
-      expect(isGranted_1).to.be.equal(true);
-      expect(isGranted_2).to.be.equal(true);
+      await condition.setAnswer(true);
+
+      // Check `pm.address` as a target `_where`
+      expect(
+        await pm.isGranted(
+          pm.address,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          condition.address
+        )
+      ).to.be.true;
+
+      // Check `address(0)` as a target `_where`
+      expect(
+        await pm.isGranted(
+          ethers.constants.AddressZero,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          condition.address
+        )
+      ).to.be.true;
+
+      // Check that `false` is returned if `otherSigner is the caller `_who`.
+      expect(
+        await pm.isGranted(
+          ethers.constants.AddressZero,
+          otherSigner.address,
+          ADMIN_PERMISSION_ID,
+          condition.address
+        )
+      ).to.be.false;
     });
 
     it('should be callable by anyone', async () => {
@@ -791,13 +867,198 @@ describe('Core: PermissionManager', function () {
         );
       expect(isGranted).to.be.equal(false);
     });
+
+    it('does not fall back to a generic caller or target condition if a specific condition is set already answering `false`', async () => {
+      const specificCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+      const genericCallerCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+      const genericTargetCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+
+      // Grant with a specific condition that will answer false
+      await pm.grantWithCondition(
+        pm.address,
+        ownerSigner.address,
+        ADMIN_PERMISSION_ID,
+        specificCondition.address
+      );
+      await specificCondition.setAnswer(false);
+
+      // Grant with a generic caller condition that will answer true
+      await pm.grantWithCondition(
+        pm.address,
+        ANY_ADDR,
+        ADMIN_PERMISSION_ID,
+        genericCallerCondition.address
+      );
+      await genericCallerCondition.setAnswer(true);
+
+      // Grant with a generic target condition that will answer true
+      await pm.grantWithCondition(
+        ANY_ADDR,
+        ownerSigner.address,
+        ADMIN_PERMISSION_ID,
+        genericTargetCondition.address
+      );
+      await genericCallerCondition.setAnswer(true);
+
+      // Check that `isGranted` returns false for `ownerSigner` to whom the specific condition was granted.
+      expect(
+        await pm.isGranted(
+          pm.address,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.false;
+
+      // Check that `ownerSigner` is still granted access to other contracts (e.g., `address(0)`) through the `genericTargetCondition` condition.
+      expect(
+        await pm.isGranted(
+          ethers.constants.AddressZero,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.true;
+    });
+
+    it('does not fall back to a generic target condition if a generic caller condition is set already answering `false`', async () => {
+      const genericCallerCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+      const genericTargetCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+
+      // Grant with a generic caller condition that will answer false.
+      await pm.grantWithCondition(
+        pm.address,
+        ANY_ADDR,
+        ADMIN_PERMISSION_ID,
+        genericCallerCondition.address
+      );
+      await genericCallerCondition.setAnswer(false);
+
+      // Grant with a generic target condition that will answer true.
+      await pm.grantWithCondition(
+        ANY_ADDR,
+        ownerSigner.address,
+        ADMIN_PERMISSION_ID,
+        genericTargetCondition.address
+      );
+      await genericTargetCondition.setAnswer(true);
+
+      // Check that `isGranted` returns false for `ANY_ADDR` (here, we check only two addresses, `ownerSigner` and `otherSigner`).
+      expect(
+        await pm.isGranted(
+          pm.address,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.false;
+      expect(
+        await pm.isGranted(
+          pm.address,
+          otherSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.false;
+
+      // Check that `ownerSigner` is granted access to other contracts (e.g., `address(0)`) via the `genericTargetCondition` condition.
+      expect(
+        await pm.isGranted(
+          ethers.constants.AddressZero,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.true;
+
+      // Check that `otherSigner` is not granted access to other contracts (e.g., `address(0)`) via the `genericTargetCondition` condition.
+      expect(
+        await pm.isGranted(
+          ethers.constants.AddressZero,
+          otherSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.false;
+    });
+
+    it('does not fall back to a generic caller or target condition if a specific condition is set already answering `false`', async () => {
+      const specificCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+      const genericCallerCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+      const genericTargetCondition = await new PermissionConditionMock__factory(
+        signers[0]
+      ).deploy();
+
+      // Grant with a specific condition that will answer false
+      await pm.grantWithCondition(
+        pm.address,
+        ownerSigner.address,
+        ADMIN_PERMISSION_ID,
+        specificCondition.address
+      );
+      await specificCondition.setAnswer(false);
+
+      // Grant with a generic caller condition that will answer true
+      await pm.grantWithCondition(
+        pm.address,
+        ANY_ADDR,
+        ADMIN_PERMISSION_ID,
+        genericCallerCondition.address
+      );
+      await genericCallerCondition.setAnswer(true);
+
+      // Grant with a generic target condition that will answer true
+      await pm.grantWithCondition(
+        ANY_ADDR,
+        ownerSigner.address,
+        ADMIN_PERMISSION_ID,
+        genericTargetCondition.address
+      );
+      await genericCallerCondition.setAnswer(true);
+
+      // Check that `isGranted` returns false for `ownerSigner` to whom the specific condition was granted.
+      expect(
+        await pm.isGranted(
+          pm.address,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.false;
+
+      // Check that `ownerSigner` is still granted access to other contracts (e.g., `address(0)`) through the `genericTargetCondition` condition.
+      expect(
+        await pm.isGranted(
+          ethers.constants.AddressZero,
+          ownerSigner.address,
+          ADMIN_PERMISSION_ID,
+          genericTargetCondition.address
+        )
+      ).to.be.true;
+    });
   });
 
   describe('_hasPermission', () => {
     let permissionCondition: PermissionConditionMock;
 
     beforeEach(async () => {
-      permissionCondition = await DeployTestPermissionCondition();
+      permissionCondition = await new PermissionConditionMock__factory(
+        ownerSigner
+      ).deploy();
     });
 
     it('should call IPermissionCondition.isGranted', async () => {